+++ aegis_gym/envs/aegis_reacher.py
@@ -55,9 +55,12 @@ class AegisReacherEnv(gym.Env):
         self.reward_type = EnvRewardType(reward_type)
 
         match self.control_type:
-            case EnvControlType.JOINTS:
+            case EnvControlType.JOINTS | EnvControlType.JOINTS_SERVO:
                 self.num_actions = 6
-            case EnvControlType.CARTESIAN_POSITION:
+            case (
+                EnvControlType.CARTESIAN_POSITION
+                | EnvControlType.CARTESIAN_POSITION_SERVO
+            ):
                 self.num_actions = 3
             case _:
                 raise ValueError(f"Unsupported control type: {self.control_type}")
@@ -116,12 +119,15 @@ class AegisReacherEnv(gym.Env):
             case EnvControlType.JOINTS:
                 dof_pos_target = self.dof_pos + delta
                 self.robot.control_dofs_position(dof_pos_target)
+            case EnvControlType.JOINTS_SERVO:
+                dof_pos_target = self.dof_pos + delta
+                self.robot.control_dofs_position_servo(dof_pos_target)
             case EnvControlType.CARTESIAN_POSITION:
                 tcp_pos_target = self.tcp_pos + delta
-                self.robot.control_tcp_position(
-                    target_pos=tcp_pos_target,
-                    target_ori=self.robot.get_tcp_orientation(),
-                )
+                self.robot.control_tcp_position(target_pos=tcp_pos_target)
+            case EnvControlType.CARTESIAN_POSITION_SERVO:
+                tcp_pos_target = self.tcp_pos + delta
+                self.robot.control_tcp_position_servo(target_pos=tcp_pos_target)
             case _:
                 raise ValueError(f"Unsupported control type: {self.control_type}")